requirements.txt
langchain-openai
langchain
python-dotenv
langchain_community
langchain_core
streamlit
langchain-groq
faiss-cpu
sentence-transformers
chromadb
pypdf